Make sure we are indeed using the new root
Apparently after move mounting new_root to /, / is still referring
to the old_root. So we first cd into new_root before move mount to
make sure we are indeed in the new_root after move mount. Then we
"chroot ." so "/" refers to the new root from now on.

Also fixed argv[0] passing